Third work of development team
we changed the model to A multi-input, multi-task hybrid model that combines DistilBERT (semantic), TF-IDF (lexical), and Category Embedding (metadata).
DistilBERT (Semantic)
Input: Product title
Purpose: Understands the meaning and relationships between words (for example, “first aid kit,” “travel,” “200-pack”).
Output: A 768-dimensional vector that represents the overall meaning of the sentence.
TF-IDF (Lexical)
Input: The same product title
Purpose: Detects explicit patterns, numbers, and abbreviations that appear frequently in titles (for example, “USB-C,” “12-pack,” “Pro+”).
Output: A text feature vector based on word frequency and two-word combinations.
Category Embedding (Metadata)
Input: The numeric ID corresponding to the product’s category
Purpose: Models the product type (for example, “Electronics,” “Home,” “Health”) as a 16-dimensional trainable vector.
Final Fusion and Outputs
The outputs of the three branches are concatenated and passed through several Dense layers to produce two outputs:
1.	Bestseller Probability (isBestSeller) → binary output (0 or 1) using a sigmoid function
2.	Price Prediction → numeric output (regression) using a linear function
When we used it on 20% of data the training took around 1 hour and half. The model was good on price prediction, but it was weak on prediction of being bestseller or not. The problems and suggestions were written on files and transported to other teammates. Also, as we had short time, team decided to change the division of data to 20% 30% and 50% to help the reduction of training time. I the pictures below this decision has been shown.